requirements.txt
pytest==8.3.5
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.10.6
faker==37.1.0
pytest-html==4.1.1
allure-pytest==2.13.5
ruff==0.11.0
black==25.1.0